refctor: github release download helper
